+++ plugins/newrelic-dashboard/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-newrelic-dashboard
 
+## 0.1.14-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.9.5-next.1
+  - @backstage/core-plugin-api@1.0.3-next.0
+  - @backstage/catalog-model@1.0.3-next.0
+  - @backstage/plugin-catalog-react@1.1.1-next.1
+
 ## 0.1.14-next.0
 
 ### Patch Changes