+++ plugins/scaffolder-backend/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage/plugin-scaffolder-backend
 
+## 1.3.0-next.1
+
+### Minor Changes
+
+- c042c5eaff: Add an option to not protect the default branch.
+
+### Patch Changes
+
+- 8f7b1835df: Updated dependency `msw` to `^0.41.0`.
+- Updated dependencies
+  - @backstage/backend-common@0.13.6-next.1
+  - @backstage/catalog-client@1.0.3-next.0
+  - @backstage/integration@1.2.1-next.1
+  - @backstage/plugin-catalog-backend@1.2.0-next.1
+  - @backstage/catalog-model@1.0.3-next.0
+  - @backstage/plugin-scaffolder-common@1.1.1-next.0
+
 ## 1.3.0-next.0
 
 ### Minor Changes